+++ packages/base/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [5.1.34]
+#### Added
+ - argv.getServerConfiguration() will resolve a httpServer.inetPub location to absolute path
+
 ## [5.1.33]
 #### Fixed
  - prevent an error in `dataLoader.localizeEntity` when running initialization scripts on entities with `softLock` mixin